pyproject.toml
[project]
name = "gaserp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.18.4",
    "dependency-injector>=4.49.0",
    "fastapi>=0.135.3",
    "passlib>=1.7.4",
    "psycopg2-binary>=2.9.11",
    "pydantic-settings>=2.13.1",
    "python-dotenv>=1.2.2",
    "python-jose>=3.5.0",
    "rich>=15.0.0",
    "sqlalchemy>=2.0.49",
    "uvicorn>=0.44.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.3",
    "ruff>=0.15.10",
]
